Put setupGraph in a helper file
Made sure that Jest ignores it as a test file (in package.json)
