+++ package.json
@@ -12,10 +12,14 @@
   },
   "jest": {
     "setupFiles": [
-      "./js/setupTests.js"
+      "<rootDir>/js/setupTests.js"
     ],
     "snapshotSerializers": [
       "enzyme-to-json/serializer"
+    ],
+    "testPathIgnorePatterns": [
+      "<rootDir>/node_modules/",
+      "<rootDir>/js/components/graph/__tests__/setupGraph.js"
     ]
   },
   "husky": {
